Reorganize code and use safer default configuration settings
